Pull upstream develop.

+++ buildscripts/create-symbols.sh
@@ -5,10 +5,12 @@ cd "$(dirname "$0")"/..
 
 # default to 0 if no build number is provided via first argument
 BUILD="${1:-0}"
+# /home/build/buildAgent/work/<some_id>/build
+BUILD_DIR="../build"
 DUMP_SYMS=./3rdparty/breakpad/src/tools/linux/dump_syms/dump_syms
 SYM_DIR=./symbols
 # library name: libglimpse.so.<version>
-LIBGLIMPSE="$(find src -type f -name 'libglimpse.so.*')"
+LIBGLIMPSE="$(find "${BUILD_DIR}" -type f -name 'libglimpse.so.*' -a ! -name '*.sym')"
 # glimpse_<version>-<build>-dbg.tar.gz
 TARBALL="glimpse_$(sed -nr 's/libglimpse\.so\.(.+)/\1/p' <<< ${LIBGLIMPSE##*/})-${BUILD}-dbg.tar.gz"
 
@@ -21,10 +23,15 @@ if [[ ! -x "${DUMP_SYMS}" ]]; then
 fi
 
 files=(
-"$(find src -type f -name glimpse-gui)"
-"$(find src -type f -name glimpse-console)"
+"$(find "${BUILD_DIR}" -type f -name glimpse-gui)"
+"$(find "${BUILD_DIR}" -type f -name glimpse-console)"
 "${LIBGLIMPSE}"
 )
+
+if [[ "${#files[*]}" -eq 0 ]]; then
+	echo "could not locate binaries" >&2
+	exit 1
+fi
 
 # remove old symbols
 rm -rf "${SYM_DIR}"
